Rename IniModuleLoader.h/.cpp to ModuleLoader.h/.cpp
This class will become the unified ModuleLoader.
