+++ srcDLL/DllMain.cpp
@@ -1,5 +1,5 @@
 #include "GameModules/IpDropDown.h"
-#include "IniModuleLoader.h"
+#include "ModuleLoader.h"
 #include "StringConversion.h"
 #include "OP2Memory.h"
 #include "FileSystemHelper.h"
