+++ srcStatic/op2extStatic.vcxproj
@@ -132,7 +132,7 @@
     <ClCompile Include="ConsoleArgumentParser.cpp" />
     <ClCompile Include="FileSystemHelper.cpp" />
     <ClCompile Include="GlobalDefines.cpp" />
-    <ClCompile Include="IniModuleLoader.cpp" />
+    <ClCompile Include="ModuleLoader.cpp" />
     <ClCompile Include="GameModules\IpDropDown.cpp" />
     <ClCompile Include="ConsoleModuleLoader.cpp" />
     <ClCompile Include="Logger.cpp" />
@@ -148,7 +148,7 @@
     <ClInclude Include="FileSystemHelper.h" />
     <ClInclude Include="GameModules\GameModule.h" />
     <ClInclude Include="GlobalDefines.h" />
-    <ClInclude Include="IniModuleLoader.h" />
+    <ClInclude Include="ModuleLoader.h" />
     <ClInclude Include="GameModules\IpDropDown.h" />
     <ClInclude Include="ConsoleModuleLoader.h" />
     <ClInclude Include="Logger.h" />
