+++ test/test.vcxproj
@@ -28,7 +28,7 @@
     <ClCompile Include="ConsoleArgumentParser.test.cpp" />
     <ClCompile Include="ConsolueModuleLoader.test.cpp" />
     <ClCompile Include="FileSystemHelper.test.cpp" />
-    <ClCompile Include="IniModuleLoader.test.cpp" />
+    <ClCompile Include="ModuleLoader.test.cpp" />
     <ClCompile Include="Logger.test.cpp" />
     <ClCompile Include="Main.cpp" />
     <ClCompile Include="op2ext.test.cpp" />
